Refactor test_backends and test_autodiff
