+++ tests/tests_geomstats/test_backends.py
@@ -860,63 +860,6 @@ class TestBackends(geomstats.tests.TestCase):
         with self.assertRaises((ValueError, RuntimeError)):
             gs.broadcast_arrays(gs.array([1, 2]), gs.array([3, 4, 5]))
 
-    @geomstats.tests.np_only
-    def test_value_and_grad_np_backend(self):
-        n = 10
-        vector = gs.ones(n)
-
-        self.assertRaises(
-            RuntimeError,
-            lambda: gs.autodiff.value_and_grad(
-                lambda v: gs.sum(v ** 2))(vector))
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_value_and_grad(self):
-        n = 10
-        vector = gs.ones(n)
-        result_loss, result_grad = gs.autodiff.value_and_grad(
-            lambda v: gs.sum(v ** 2))(vector)
-        expected_loss = n
-        expected_grad = 2 * vector
-
-        self.assertAllClose(result_loss, expected_loss)
-        self.assertAllClose(result_grad, expected_grad)
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_value_and_grad_numpy_input(self):
-        n = 10
-        vector = _np.ones(n)
-        result_loss, result_grad = gs.autodiff.value_and_grad(
-            lambda v: gs.sum(v ** 2))(vector)
-        expected_loss = n
-        expected_grad = 2 * vector
-        self.assertAllClose(result_loss, expected_loss)
-        self.assertAllClose(result_grad, expected_grad)
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_value_and_grad_two_vector_vars(self):
-        def func(x, y):
-            return gs.sum((x - y) ** 2)
-
-        arg_x, arg_y = 1., 2.
-        val, grad = gs.autodiff.value_and_grad(func)(arg_x, arg_y)
-        self.assertAllClose(val, 1.)
-        self.assertAllClose(grad, (-2., 2.))
-
-        arg_x = gs.array([1., 2.])
-        arg_y = gs.array([2., 3.])
-        val, grad = gs.autodiff.value_and_grad(func)(arg_x, arg_y)
-        self.assertAllClose(val, 2.)
-        self.assertAllClose(grad[0], gs.array([-2., -2.]))
-        self.assertAllClose(grad[1], gs.array([2., 2.]))
-
-        arg_x = gs.array([[1., 3.], [2., 3.]])
-        arg_y = gs.array([[2., 5.], [0., 4.]])
-        val, grad = gs.autodiff.value_and_grad(func)(arg_x, arg_y)
-        self.assertAllClose(val, 10.)
-        self.assertAllClose(grad[0], gs.array([[-2., -4.], [4., -2.]]))
-        self.assertAllClose(grad[1], gs.array([[2., 4.], [-4., 2.]]))
-
     def test_choice(self):
         x = gs.array([0.1, 0.2, 0.3, 0.4, 0.5])
         a = 4
@@ -1017,30 +960,6 @@ class TestBackends(geomstats.tests.TestCase):
         expected = _np.linalg.cholesky(mat)
         self.assertAllClose(result, expected)
 
-    @geomstats.tests.autograd_and_torch_only
-    def test_expm_backward(self):
-        mat = gs.array([[0, 1, .5], [-1, 0, 0.2], [-.5, -.2, 0]])
-        mat = gs.cast(mat, gs.float64)
-
-        def loss(p):
-            return gs.sum((gs.linalg.expm(p) - gs.eye(3)) ** 2)
-
-        value_and_grad = gs.autodiff.value_and_grad(loss)
-        result = value_and_grad(mat)
-
-        def loss_torch(p):
-            return torch.sum((torch.matrix_exp(p) - torch.eye(3)) ** 2)
-
-        torch_mat = torch.tensor(
-            [[0, 1, .5], [-1, 0, 0.2], [-.5, -.2, 0]], dtype=torch.float64,
-            requires_grad=True)
-        value = loss_torch(torch_mat)
-        value.backward()
-        grad = torch_mat.grad
-
-        self.assertAllClose(result[0], value.detach())
-        self.assertAllClose(result[1], grad)
-
     def test_mat_from_diag_triu_tril(self):
 
         diag = gs.array([9., 9., 9.])
@@ -1071,394 +990,3 @@ class TestBackends(geomstats.tests.TestCase):
         result = gs.prod(vec)
         expected = gs.cumprod(vec)[-1]
         self.assertAllClose(result, expected)
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_custom_grad_one_var(self):
-
-        def grad_x(x):
-            return 2 * x
-
-        @gs.autodiff.custom_gradient(grad_x)
-        def func(x):
-            return gs.sum(x ** 2)
-
-        arg_x = gs.array([1., 3.])
-        result_value, result_grad = gs.autodiff.value_and_grad(func)(arg_x)
-
-        expected_value = 10.
-        expected_grad = gs.array([2., 6.])
-
-        self.assertAllClose(result_value, expected_value)
-        self.assertAllClose(result_grad, expected_grad)
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_custom_grad_two_vars(self):
-
-        def grad_x(x, y):
-            return 2 * (x - y)
-
-        def grad_y(x, y):
-            return 2 * (y - x)
-
-        @gs.autodiff.custom_gradient(grad_x, grad_y)
-        def func(x, y):
-            return gs.sum((x - y) ** 2)
-
-        arg_x = gs.array([[1., 3.], [2., 3.]])
-        arg_y = gs.array([[2., 5.], [0., 4.]])
-
-        result_val, result_grad = gs.autodiff.value_and_grad(func)(
-            arg_x, arg_y)
-
-        self.assertTrue(isinstance(result_grad, tuple))
-        result_grad_x, result_grad_y = result_grad
-
-        expected_val = func(arg_x, arg_y)
-        expected_grad_x = grad_x(arg_x, arg_y)
-        expected_grad_y = grad_y(arg_x, arg_y)
-
-        self.assertAllClose(result_val, expected_val)
-        self.assertAllClose(result_grad_x, expected_grad_x)
-        self.assertAllClose(result_grad_y, expected_grad_y)
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_custom_grad_two_vars_madeup_grads(self):
-
-        def grad_x(x, y):
-            return 6 * (x - y)
-
-        def grad_y(x, y):
-            return 6 * (y - x)
-
-        @gs.autodiff.custom_gradient(grad_x, grad_y)
-        def func(x, y):
-            return gs.sum((x - y) ** 2)
-
-        arg_x = gs.array([[1., 3.], [2., 3.]])
-        arg_y = gs.array([[2., 5.], [0., 4.]])
-
-        result_val, result_grad = gs.autodiff.value_and_grad(func)(
-            arg_x, arg_y)
-
-        self.assertTrue(isinstance(result_grad, tuple))
-        result_grad_x, result_grad_y = result_grad
-
-        expected_val = func(arg_x, arg_y)
-        expected_grad_x = grad_x(arg_x, arg_y)
-        expected_grad_y = grad_y(arg_x, arg_y)
-
-        self.assertAllClose(result_val, expected_val)
-        self.assertAllClose(result_grad_y, expected_grad_y)
-        self.assertAllClose(result_grad_x, expected_grad_x)
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_custom_grad_dummy_two_vars_composed(self):
-
-        def grad_x(x, y):
-            return 2 * (x - y)
-
-        def grad_y(x, y):
-            return 2 * (y - x)
-
-        @gs.autodiff.custom_gradient(grad_x, grad_y)
-        def func(x, y):
-            return gs.sum((x - y) ** 2)
-
-        arg_x = gs.array([[1., 3.], [2., 3.]])
-        arg_y = gs.array([[2., 5.], [0., 4.]])
-
-        def func_2(x):
-            return gs.exp(-.5 * func(x, arg_y))
-
-        result_value, result_grad = gs.autodiff.value_and_grad(func_2)(arg_x)
-        expected_value = func_2(arg_x)
-        expected_grad = (arg_y - arg_x) * expected_value
-        self.assertAllClose(result_value, expected_value)
-        self.assertAllClose(result_grad, expected_grad)
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_custom_grad_dummy_two_vars_composed_madeup_grad(self):
-
-        def grad_x(x, y):
-            return 6 * (x - y)
-
-        def grad_y(x, y):
-            return 6 * (y - x)
-
-        @gs.autodiff.custom_gradient(grad_x, grad_y)
-        def func(x, y):
-            return gs.sum((x - y) ** 2)
-
-        arg_x = gs.array([[1., 3.], [2., 3.]])
-        arg_y = gs.array([[2., 5.], [0., 4.]])
-
-        def func_2(x):
-            return gs.exp(-.5 * func(x, arg_y))
-
-        result_value, result_grad = gs.autodiff.value_and_grad(func_2)(arg_x)
-
-        expected_value = func_2(arg_x)
-        expected_grad = 3 * (arg_y - arg_x) * expected_value
-
-        self.assertAllClose(result_value, expected_value)
-        self.assertAllClose(result_grad, expected_grad)
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_custom_grad_unused_var(self):
-
-        def grad_dummy(dummy, x, y):
-            return dummy
-
-        def grad_x(dummy, x, y):
-            return 2 * dummy * (x - y)
-
-        def grad_y(dummy, x, y):
-            return 2 * dummy * (y - x)
-
-        @gs.autodiff.custom_gradient(grad_dummy, grad_x, grad_y)
-        def func(dummy, x, y):
-            return dummy * gs.sum((x - y) ** 2)
-
-        const_y = gs.array([[2., 5.], [0., 4.]])
-        const_dummy = gs.array(4.)
-
-        def func_of_x(x):
-            return func(const_dummy, x, const_y)
-
-        arg_x = gs.array([[1., 3.], [2., 3.]])
-        result_value, result_grad = gs.autodiff.value_and_grad(
-            func_of_x)(arg_x)
-        expected_value = func_of_x(arg_x)
-        expected_grad = grad_x(const_dummy, arg_x, const_y)
-
-        self.assertAllClose(result_value, expected_value)
-        self.assertAllClose(result_grad, expected_grad)
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_custom_grad_unused_var_madeup_grad(self):
-
-        def grad_dummy(dummy, x, y):
-            return dummy
-
-        def grad_x(dummy, x, y):
-            return 6 * dummy * (x - y)
-
-        def grad_y(dummy, x, y):
-            return 6 * dummy * (y - x)
-
-        @gs.autodiff.custom_gradient(grad_dummy, grad_x, grad_y)
-        def func(dummy, x, y):
-            return dummy * gs.sum((x - y) ** 2)
-
-        const_y = gs.array([[2., 5.], [0., 4.]])
-        const_dummy = gs.array(4.)
-
-        def func_of_x(x):
-            return func(const_dummy, x, const_y)
-
-        arg_x = gs.array([[1., 3.], [2., 3.]])
-        result_value, result_grad = gs.autodiff.value_and_grad(
-            func_of_x)(arg_x)
-        expected_value = func_of_x(arg_x)
-        expected_grad = grad_x(const_dummy, arg_x, const_y)
-
-        self.assertAllClose(result_value, expected_value)
-        self.assertAllClose(result_grad, expected_grad)
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_custom_grad_unused_var_chain_rule(self):
-
-        def grad_dummy(dummy, x, y):
-            return dummy
-
-        def grad_x(dummy, x, y):
-            return 6 * dummy * (x - y)
-
-        def grad_y(dummy, x, y):
-            return 6 * dummy * (y - x)
-
-        @gs.autodiff.custom_gradient(grad_dummy, grad_x, grad_y)
-        def func(dummy, x, y):
-            return dummy * gs.sum((x - y) ** 2)
-
-        const_y = gs.array([[2., 5.], [0., 4.]])
-        const_dummy = gs.array(4.)
-
-        def func_2(x):
-            return gs.exp(-.5 * func(const_dummy, x, const_y))
-
-        arg_x = gs.array([[1., 3.], [2., 3.]])
-        result_value, result_grad = gs.autodiff.value_and_grad(
-            func_2)(arg_x)
-        expected_value = func_2(arg_x)
-        expected_grad = - .5 * grad_x(
-            const_dummy, arg_x, const_y) * func_2(arg_x)
-
-        self.assertAllClose(result_value, expected_value)
-        self.assertAllClose(result_grad, expected_grad)
-
-    @geomstats.tests.autograd_and_tf_only
-    def test_custom_grad_mock_squared_dist(self):
-
-        def squared_dist_grad_a(point_a, point_b, metric):
-            return 2 * metric.log(point_b, point_a)
-
-        def squared_dist_grad_b(point_a, point_b, metric):
-            return 2 * metric.log(point_a, point_b)
-
-        @gs.autodiff.custom_gradient(
-            squared_dist_grad_a, squared_dist_grad_b)
-        def squared_dist(point_a, point_b, metric):
-            dist = metric.squared_dist(point_a, point_b)
-            return dist
-
-        space = SpecialEuclidean(n=2)
-        const_metric = space.left_canonical_metric
-        const_point_b = space.random_point()
-
-        def func(x):
-            return squared_dist(x, const_point_b, metric=const_metric)
-
-        arg_point_a = space.random_point()
-        expected_value = func(arg_point_a)
-        expected_grad = 2 * const_metric.log(const_point_b, arg_point_a)
-        result_value, result_grad = gs.autodiff.value_and_grad(
-            func)(arg_point_a)
-        self.assertAllClose(result_value, expected_value)
-        self.assertAllClose(result_grad, expected_grad)
-
-    @geomstats.tests.autograd_and_tf_only
-    def test_custom_grad_in_action(self):
-        space = SpecialEuclidean(n=2)
-        const_metric = space.left_canonical_metric
-        const_point_b = space.random_point()
-
-        def func(x):
-            return const_metric.squared_dist(x, const_point_b)
-
-        arg_point_a = space.random_point()
-        result_value, result_grad = gs.autodiff.value_and_grad(
-            func)(arg_point_a)
-        expected_value = const_metric.squared_dist(
-            arg_point_a, const_point_b)
-        expected_grad = 2 * const_metric.log(
-            const_point_b, arg_point_a)
-
-        self.assertAllClose(result_value, expected_value)
-        self.assertAllClose(result_grad, expected_grad)
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_custom_grad_chain_rule_one_var(self):
-
-        def fun1_grad(x):
-            return 1.
-
-        @gs.autodiff.custom_gradient(fun1_grad)
-        def fun1(x):
-            return x
-
-        def fun2(x):
-            out = fun1(x) ** 2
-            return out
-
-        def fun2_grad(x):
-            return 2 * x
-
-        arg = gs.array(10.)
-
-        result_value, result_grad = gs.autodiff.value_and_grad(fun2)(arg)
-        expected_value = fun2(arg)
-        expected_grad = 20.
-
-        expected_grad2 = fun2_grad(fun1(arg)) * fun1_grad(arg)
-
-        self.assertAllClose(result_value, expected_value)
-        self.assertAllClose(result_grad, expected_grad)
-        self.assertAllClose(result_grad, expected_grad2)
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_custom_grad_chain_rule_one_var_madeup_grad(self):
-
-        def fun1_grad(x):
-            return 3.
-
-        @gs.autodiff.custom_gradient(fun1_grad)
-        def fun1(x):
-            return x
-
-        def fun2(x):
-            out = fun1(x) ** 2
-            return out
-
-        def fun2_grad(x):
-            return 2 * x
-
-        arg = gs.array(10.)
-
-        result_value, result_grad = gs.autodiff.value_and_grad(fun2)(arg)
-        expected_value = fun2(arg)
-        expected_grad_explicit = 60.
-
-        expected_grad_chain_rule = fun2_grad(fun1(arg)) * fun1_grad(arg)
-
-        self.assertAllClose(result_value, expected_value)
-        self.assertAllClose(result_grad, expected_grad_explicit)
-        self.assertAllClose(result_grad, expected_grad_chain_rule)
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_custom_grad_chain_rule_one_vector_var(self):
-
-        def fun1_grad(x):
-            return 2 * x
-
-        @gs.autodiff.custom_gradient(fun1_grad)
-        def fun1(x):
-            return gs.sum(x) ** 2
-
-        def fun2(x):
-            out = fun1(x) ** 3
-            return out
-
-        def fun2_grad(x):
-            return 3 * x ** 2
-
-        arg = gs.array([1., 2.])
-
-        result_value, result_grad = gs.autodiff.value_and_grad(fun2)(arg)
-
-        expected_value = fun2(arg)
-        expected_grad_explicit = 6 * fun1(arg) ** 2 * arg
-        expected_grad_chain_rule = fun2_grad(fun1(arg)) * fun1_grad(arg)
-
-        self.assertAllClose(result_value, expected_value)
-        self.assertAllClose(result_grad, expected_grad_explicit)
-        self.assertAllClose(result_grad, expected_grad_chain_rule)
-
-    @geomstats.tests.autograd_tf_and_torch_only
-    def test_custom_grad_chain_rule_one_vector_var_madeup_grad(self):
-
-        def fun1_grad(x):
-            return 6 * x
-
-        @gs.autodiff.custom_gradient(fun1_grad)
-        def fun1(x):
-            return gs.sum(x) ** 2
-
-        def fun2(x):
-            out = fun1(x) ** 3
-            return out
-
-        def fun2_grad(x):
-            return 3 * x ** 2
-
-        arg = gs.array([1., 2.])
-
-        result_value, result_grad = gs.autodiff.value_and_grad(fun2)(arg)
-
-        expected_value = fun2(arg)
-        expected_grad_explicit = 18 * fun1(arg) ** 2 * arg
-        expected_grad_chain_rule = fun2_grad(fun1(arg)) * fun1_grad(arg)
-
-        self.assertAllClose(result_value, expected_value)
-        self.assertAllClose(result_grad, expected_grad_explicit)
-        self.assertAllClose(result_grad, expected_grad_chain_rule)
